[DC-2111] Rename file ct
